% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehr2pheno.R
\name{ehr2pheno}
\alias{ehr2pheno}
\title{Phenotyping from EHR data}
\usage{
ehr2pheno(codes, nlp, ...)
}
\arguments{
\item{codes}{a \code{n x q} matrix containg the occurences of the \code{q} diagnosis
codes related to the phenotype of interest for all \code{n} patients}

\item{nlp}{a \code{n x p} matrix containg the occurences of the \code{q} diagnosis
codes related to the phenotype of interest for all \code{n} patients}

\item{\dots}{further arguments to be passed to \code{\link{get_thresholds}} or \code{\link{phen_feat_sel}}
subfunctions}
}
\value{
an object of type \code{ehr2pheno}, which is a list of the following elements:
\itemize{
 \item{\code{feat_sel}: }{itself a list with the three \code{glmnet} models,
   \code{"codes"}, \code{"nlp"} and  \code{"both"}}
 \item{\code{surrogates}: }{itself a list with the three constructed surrogates,
   \code{"codes"}, \code{"nlp"} and  \code{"both"}}
}
}
\description{
Using both diagnosis codes and CUIs extracted from NLP processed notes and selecting
predicting features
}
