% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_thresholds.R
\name{get_thresholds}
\alias{get_thresholds}
\title{Split surrogate into 3 ordered categories based on the number of feature occurences}
\usage{
get_thresholds(x, method = c("quantiles", "GaussianMixture"), p = 0.1,
  ...)
}
\arguments{
\item{x}{a feature occurence count vector to be split into 3 components}

\item{method}{a character string indicating which splitting method is
used : either \code{"GaussianMixture"} or \code{"quantiles"}. Default is
\code{"quantiles"}.}

\item{p}{the probability of extemes used in the quantile splitting method
(ignored if \code{method} is \code{"GaussianMixture"}). Default is \code{0.01}.}

\item{\dots}{further agruments not used here}
}
\value{
a list with the following components:
 \itemize{
 \item{\code{thesholds}}{a list containing the lower threshold \code{lowthres}
 marking the difference between the two classes \code{0} (the fewest occurences) and \code{0.5};
 and the upper threshold \code{upthres} marking the difference between the
 two classes \code{0.5} and \code{1} (the most occurences)}
 \item{surrogate} a vector containing the clustering for each observation (as
 an ordered factor with the 3 following levels ordered by the number of occurences:
 \code{0 < 0.5 < 1}.)
 }
}
\description{
Either fit 3 components Gaussian mixture to determine cutpoints for surrogate variable
of uses the most extremes quantiles
}
