% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_feat_sel.R
\name{pheno_feat_sel}
\alias{pheno_feat_sel}
\title{Feature selection}
\usage{
pheno_feat_sel(surrogate, features, ...)
}
\arguments{
\item{surrogate}{a vector of length \code{n} containing. Coerced to numeric.}

\item{features}{a matrix of dimension \code{n} x \code{p} containing the features
from which to select predictors}

\item{\dots}{further agruments not used here}
}
\value{
a list containing 2 components:
\itemize{
\item{logisticEN_model}{ a \code{\link[glmnet]{glmnet}} object
that can be used to predict external observation.}
\item{sel_feat}{a vector containing either the names of the selected features
via cross valiation on the "class" measure type (through \code{\link[glmnet]{cv.glmnet}})
using the \code{lambda.1se} parameter.}
}
}
\description{
Elastic net logistic regression
}
