% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.summary.ehr2pheno.R
\name{plot.summary.ehr2pheno}
\alias{plot.summary.ehr2pheno}
\title{Plotting ROC curves from the summary of an ehr2pheno Object}
\usage{
\method{plot}{summary.ehr2pheno}(x, ...)
}
\arguments{
\item{x}{\code{summary.ehr2pheno} object}

\item{\dots}{Arguments to be passed to or from other methods}
}
\description{
Plotting ROC curves from the summary of an ehr2pheno Object
}
\details{
WARNING: the ROC are computed on the training data itself and
are therefore over-optimistic !
}
\seealso{
\code{\link{ehr2pheno}} \code{\link{summary.ehr2pheno}}
}
