% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ehr2pheno.R
\name{summary.ehr2pheno}
\alias{summary.ehr2pheno}
\title{Summarize a phenotype prediction}
\usage{
\method{summary}{ehr2pheno}(object, goldstandard = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{"ehr2pheno"}}

\item{goldstandard}{a binary vector containing the gold-standard phenotypes}

\item{\dots}{Arguments to be passed to or from other methods}
}
\value{
an object of class \code{"summary.ehr2pheno"} with the following elements:
\itemize{
 \item{\code{criteria}: }{\itemize{
   \item{\code{AUC}: }{a vector of length 3 containing the AUC on the training data for
     \code{codes} only, \code{nlp} only, and \code{both} codes and nlp.}
   \item{\code{TPR}: }{a matrix of 3 columns the TPR for each threshold
     on the training data for \code{codes} only, \code{nlp} only, and
     \code{both} codes and nlp.}
   \item{\code{FPR}: }{a matrix of 3 columns the FPR for each threshold
     on the training data for \code{codes} only, \code{nlp} only, and
     \code{both} codes and nlp.}
 }}
 \item{\code{pred_proba}: }{a list with the predicted probabilities for each of the
 three models with \code{codes}, \code{nlp} and \code{both}}
 \item{\code{pred_proba}: }{a list with the selected features and associated odds ratios for each of the
 three logistic regression models with elastic-net \code{codes}, \code{nlp} and \code{both}}
}
}
\description{
Summarize a phenotype prediction from diagnosis codes and NLP processed notes
}
\details{
Computes the True Positives Rate (TPR), False Positives Rate (FPR),
and Area Under the ROC Curve (AUC).

WARNING: the criteria are computed on the training data itself and
are therefore over-optimistic !
}
\seealso{
\code{\link{ehr2pheno}}
}
